refactor: remove duplicate cn, organize utils
